requirements.txt
pyyaml
requests
wcwidth
